fix: move introspection cache lookup to own file
