+++ src/lib/idp/introspectToken.ts
@@ -4,7 +4,7 @@ import axios, { type AxiosError } from "axios";
 import { createPrivateKey } from "node:crypto";
 import { ZITADEL_APPLICATION_KEY, ZITADEL_DOMAIN } from "@src/config";
 
-type IntrospectionResult = {
+export type IntrospectionResult = {
   username: string;
   exp: number;
 };
