+++ src/middleware/authentication/middleware.ts
@@ -1,15 +1,9 @@
-// biome-ignore lint/correctness/noNodejsModules: we need the node crypto module
-import { createHash } from "node:crypto";
 import type { NextFunction, Request, Response } from "express";
+import {
+  getIntrospectionCache,
+  setIntrospectionCache,
+} from "@lib/idp/introspectionCache";
 import { introspectToken } from "@lib/idp/introspectToken";
-import { RedisConnector } from "@src/lib/redisConnector";
-
-const cacheExpiry = 300; // seconds
-
-export function getAccessTokenCacheKey(accessToken: string): string {
-  const hash = createHash("sha256").update(accessToken).digest("base64");
-  return `api:auth:${hash}`;
-}
 
 export async function authenticationMiddleware(
   request: Request,
@@ -17,21 +11,14 @@ export async function authenticationMiddleware(
   next: NextFunction,
 ) {
   const accessToken = request.headers.authorization?.split(" ")[1];
-  const redisConnector = await RedisConnector.getInstance();
 
   if (!accessToken) {
     return response.sendStatus(401);
   }
 
-  // Check if there is a cached introspection result for this token.
-  // This is being done to reduce strain on the IdP.
-  const accessTokenKey = getAccessTokenCacheKey(accessToken);
-  const introspectionCached = await redisConnector.client.get(accessTokenKey);
-  console.debug("Introspection cache", accessTokenKey, introspectionCached);
-
-  const introspectionResult = introspectionCached
-    ? JSON.parse(introspectionCached)
-    : await introspectToken(accessToken);
+  const introspectionResult =
+    (await getIntrospectionCache(accessToken)) ??
+    (await introspectToken(accessToken));
 
   if (!introspectionResult) {
     return response.sendStatus(403);
@@ -47,12 +34,7 @@ export async function authenticationMiddleware(
     return response.status(401).json({ message: "Token expired" });
   }
 
-  // Store the introspection result in the cache
-  await redisConnector.client.set(
-    accessTokenKey,
-    JSON.stringify(introspectionResult),
-    { EX: cacheExpiry },
-  );
+  await setIntrospectionCache(accessToken, introspectionResult);
 
   next();
 }
